requirements.txt
vonage~=4.0
fastapi
uvicorn[standard]==0.32.0
python-dotenv==1.0.1
pydantic
requests
vonage_cloud_runtime
bson
motor
